Unique classnames across test directories
